Implement ExtensionIndex
